pyproject.toml
[project]
name = "bigquery-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.26.0",
    "ruff>=0.15.5",
]
